+++ php/Terminus/Models/TerminusModel.php
@@ -7,15 +7,26 @@ use Terminus\Exceptions\TerminusException;
 use Terminus\Request;
 
 abstract class TerminusModel {
+  /**
+   * @var array Arguments for fetching this model's information
+   */
+  protected $args = [];
+  /**
+   * @var object
+   */
+  protected $attributes;
+  /**
+   * @var string
+   */
   protected $id;
   /**
    * @var Request
    */
   protected $request;
   /**
-   * @var object
+   * @var string The URL at which to fetch this model's information
    */
-  protected $attributes;
+  protected $url;
 
   /**
    * Object constructor
@@ -79,14 +90,16 @@ abstract class TerminusModel {
 
     $options = array_merge(
       array('options' => array('method' => 'get')),
-      $this->getFetchArgs(),
+      $this->args,
       $fetch_args
     );
 
-    $results = $this->request->request(
-      $this->getFetchUrl(),
-      $options
-    );
+    if (isset($this->url)) {
+      $fetch_url = $this->url;
+    } else {
+      $fetch_url = $this->getFetchUrl();
+    }
+    $results = $this->request->request($fetch_url, $options);
 
     $data = $results['data'];
     $data = $this->parseAttributes($data);
@@ -127,15 +140,6 @@ abstract class TerminusModel {
   public function has($attribute) {
     $isset = isset($this->attributes->$attribute);
     return $isset;
-  }
-
-  /**
-   * Give necessary args for collection data fetching
-   *
-   * @return array
-   */
-  protected function getFetchArgs() {
-    return array();
   }
 
   /**